[test] help MSVC to deduce correct type

+++ tests/nsl/geom/NSLGeomTest.cpp
@@ -179,7 +179,7 @@ void NSLGeomTest::testLineSim() {
 	QCOMPARE(aerr, 0.2);
 
 	// nout can be up to 10 when third arg of nsl_geom_linesim_visvalingam_whyatt() is 10
-	for (i = 0; i < std::min(nout, 6uL); ++i)
+	for (i = 0; i < std::min(nout, (size_t)6uL); ++i)
 		QCOMPARE(index[i], result3[i]);
 
 	const size_t result5[] = {0, 2, 3, 5, 6, 7, 9};
